requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
distlib==0.3.4
filelock==3.7.1
Flask==2.1.2
fonttools==4.33.3
gunicorn==20.1.0
h11==0.13.0
idna==3.3
imageio==2.19.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.1
networkx==2.8.4
numpy==1.23.0
oauthlib==3.2.0
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.0
platformdirs==2.5.2
pydantic==1.9.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.3.0
requests==2.27.1
requests-oauthlib==1.3.0
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.1
six==1.16.0
threadpoolctl==3.1.0
tifffile==2022.5.4
typing-extensions==4.2.0
urllib3==1.26.8
uvicorn==0.17.0.post1
virtualenv==20.13.0
Werkzeug==2.1.2
zipp==3.8.0
flask_cors==3.0.10

